+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/ui/components/spl/SoundLevelMeterViewModel.kt
@@ -6,7 +6,7 @@ import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
 import org.noiseplanet.noisecapture.services.liveaudio.LiveAudioService
 
-class SPLIndicatorsViewModel : ViewModel(), KoinComponent {
+class SoundLevelMeterViewModel : ViewModel(), KoinComponent {
 
     // - Constants
