+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/ui/features/measurement/MeasurementModule.kt
@@ -2,7 +2,7 @@ package org.noiseplanet.noisecapture.ui.features.measurement
 
 import org.koin.core.module.dsl.viewModel
 import org.koin.dsl.module
-import org.noiseplanet.noisecapture.ui.components.spl.SPLIndicatorsViewModel
+import org.noiseplanet.noisecapture.ui.components.spl.SoundLevelMeterViewModel
 import org.noiseplanet.noisecapture.ui.features.measurement.plot.spectrogram.SpectrogramPlotViewModel
 import org.noiseplanet.noisecapture.ui.features.measurement.plot.spectrum.SpectrumPlotViewModel
 
@@ -10,7 +10,7 @@ import org.noiseplanet.noisecapture.ui.features.measurement.plot.spectrum.Spectr
 val measurementModule = module {
 
     viewModel {
-        SPLIndicatorsViewModel()
+        SoundLevelMeterViewModel()
     }
 
     viewModel {
