+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/ui/features/measurement/MeasurementScreen.kt
@@ -19,7 +19,7 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleEventObserver
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.compose.LocalLifecycleOwner
-import org.noiseplanet.noisecapture.ui.components.spl.SPLIndicatorsView
+import org.noiseplanet.noisecapture.ui.components.spl.SoundLevelMeterView
 import org.noiseplanet.noisecapture.ui.features.measurement.controls.RecordingControls
 
 const val DEFAULT_SAMPLE_RATE = 48000.0
@@ -77,7 +77,7 @@ fun MeasurementScreen(
             if (maxWidth > maxHeight) {
                 Row(modifier = Modifier.fillMaxSize()) {
                     Column(modifier = Modifier.fillMaxWidth(.5F)) {
-                        SPLIndicatorsView(viewModel = viewModel.splIndicatorsViewModel)
+                        SoundLevelMeterView(viewModel = viewModel.soundLevelMeterViewModel)
                         RecordingControls(viewModel = viewModel.recordingControlsViewModel)
                     }
                     Column(modifier = Modifier) {
@@ -86,7 +86,7 @@ fun MeasurementScreen(
                 }
             } else {
                 Column(modifier = Modifier.fillMaxSize()) {
-                    SPLIndicatorsView(viewModel = viewModel.splIndicatorsViewModel)
+                    SoundLevelMeterView(viewModel = viewModel.soundLevelMeterViewModel)
                     RecordingControls(
                         viewModel = viewModel.recordingControlsViewModel,
                         modifier = Modifier.height(IntrinsicSize.Min)
